Allow for Stock Bookmark Removal from Any Page
